Dec alphaserver patch from Malcolm Beattie

+++ src/backend/libpq/pqcomprim.c
@@ -27,7 +27,7 @@
 #if BYTE_ORDER == BIG_ENDIAN
 #define ntoh_s(n) (u_short)(((u_char *)&n)[1] << 8 \
 							  | ((u_char *)&n)[0])
-#define ntoh_l(n) (u_long) (((u_char *)&n)[3] << 24 \
+#define ntoh_l(n) (uint32) (((u_char *)&n)[3] << 24 \
 							  | ((u_char *)&n)[2] << 16 \
 							  | ((u_char *)&n)[1] <<  8 \
 							  | ((u_char *)&n)[0])
@@ -65,10 +65,10 @@ int
 pqPutLong(int integer, FILE *f)
 {
 	int			retval = 0;
-	u_long		n;
+	uint32		n;
 
 	n = hton_l(integer);
-	if (fwrite(&n, sizeof(u_long), 1, f) != 1)
+	if (fwrite(&n, sizeof(uint32), 1, f) != 1)
 		retval = EOF;
 
 	return retval;
@@ -93,9 +93,9 @@ int
 pqGetLong(int *result, FILE *f)
 {
 	int			retval = 0;
-	u_long		n;
+	uint32		n;
 
-	if (fread(&n, sizeof(u_long), 1, f) != 1)
+	if (fread(&n, sizeof(uint32), 1, f) != 1)
 		retval = EOF;
 
 	*result = ntoh_l(n);
